+++ OpenMarket/OpenMarket/View/ProductGridCell.swift
@@ -85,7 +85,7 @@ class ProductGridCell: UICollectionViewCell {
         productThumbnail.image = UIImage(data: thumbnailData)
     }
     
-    private func setPriceLabel(_ currency: String, _ price: Double, _ discountedPrice: Double) {
+    private func setPriceLabel(_ currency: String, _ price: Int, _ discountedPrice: Int) {
         let numberFormatter = NumberFormatter()
         numberFormatter.numberStyle = .decimal
         guard let formattedPrice = numberFormatter.string(from: NSNumber(value: price)) else { return }
